optimized point extraction
