+++ scripts/data-processing/requirements.txt
@@ -1,0 +1,3 @@
+pandas
+pdfplumber
+pyyaml
